requirements.txt
pandas==2.2.2
scikit-learn==1.5.1
matplotlib==3.9.2
shap==0.46.0
networkx==3.3
igraph==0.11.6
seaborn==0.13.2
xgboost==2.1.1
tabulate==0.9.0
utils==1.0.2
ipython==8.27.0
fpdf==1.7.2
python-dotenv==1.0.1